refactor images lab so that it becomes easy to extend new styles
