+++ PowerPointLabs/PowerPointLabs/ImagesLab/Factory/StyleVariantsFactory.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
+using PowerPointLabs.ImagesLab.Domain;
 
-namespace PowerPointLabs.ImagesLab.Domain
+namespace PowerPointLabs.ImagesLab.Factory
 {
     class StyleVariantsFactory
     {
@@ -840,53 +841,6 @@ namespace PowerPointLabs.ImagesLab.Domain
                     {"OptionName", "Font Size +21"},
                     {"IsUseTextFormat", true},
                     {"FontSizeIncrease", 21}
-                })
-            };
-        }
-
-        private static List<StyleVariants> GetImageOffsetVariants()
-        {
-            return new List<StyleVariants>
-            {
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "No offset"},
-                    {"ImageOffset", 0}
-                }),
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "Offset +20%"},
-                    {"ImageOffset", 20}
-                }),
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "Offset +50%"},
-                    {"ImageOffset", 50}
-                }),
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "Offset +80%"},
-                    {"ImageOffset", 80}
-                }),
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "Offset -20%"},
-                    {"ImageOffset", -20}
-                }),
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "Offset -50%"},
-                    {"ImageOffset", -50}
-                }),
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "Offset -80%"},
-                    {"ImageOffset", -80}
-                }),
-                new StyleVariants(new Dictionary<string, object>
-                {
-                    {"OptionName", "Offset -100%"},
-                    {"ImageOffset", -100}
                 })
             };
         }
